==> 1_initial_analysis.Rmd <==
---
title: "1_initial_analysis"
author: "Ajay Prakash Nair"
date: "2024-10-24"
output: html_document
---


```{r setup, include=FALSE}
# Set CRAN mirror
options(repos = c(CRAN = "https://cran.rstudio.com/"))

# Other setup configurations
knitr::opts_chunk$set(echo = TRUE)
```

# Requirements

```{r requirements}
requirements=c("tidyverse","mice", "caTools", "corrplot", "summarytools", "plotly", "readr", "caret")

for (req in requirements){
  if (!require(req, character.only = TRUE)){
      install.packages(req)
  }
}
```

# Introduction

The objective of this project is to analyze the statistical data of the Spanish La Liga football league spanning the last 9 seasons and predicting the result for the 2023/2024 season. The dataset, sourced from http://www.football-data.co.uk/, provides comprehensive information on various aspects of each match, including final and half-time results, corner kicks, and disciplinary actions such as yellow and red cards. This dataset serves as a valuable resource for understanding the dynamics of football matches in one of Europe's top football leagues.

# Data Description

The dataset comprises detailed statistical records of matches played in the Spanish La Liga over the past decade. Each record includes information such as match date, teams involved, final and half-time scores, number of corner kicks, as well as disciplinary actions like yellow and red cards.

The different information of each match collected on the dataset is described in the following table:

| **Label** | **Description** |
|-----|:-------------|
|Date| Date of the match|
|HomeTeam| Home Team of the match|
|AwayTeam| Away Team of the match|
|FTHG| Full Time Home Team Goals|
|FTAG| Full Time Away Team Goals|
|FTR| Full Time Result (H=Home Win, D=Draw, A=Away Win)|
|HTHG| Half Time Home Team Goals|
|HTAG| Half Time Away Team Goals|
|HTR| Half Time Result (H=Home Win, D=Draw, A=Away Win)|
|HS| Home Team Shots|
|AS| Away Team Shots|
|HST| Home Team Shots on Target|
|AST| Away Team Shots on Target|
|HF| Home Team Fouls Committed|
|AF| Away Team Fouls Committed|
|HC| Home Team Corners|
|AC| Away Team Corners|
|HY| Home Team Yellow Cards|
|AY| Away Team Yellow Cards|
|HR| Home Team Red Cards|
|AR| Away Team Red Cards|


## Analysis description

- To identify trends and patterns in match outcomes over the past 9 seasons of La Liga.
- To explore the impact of various factors such as home advantage, team form, and disciplinary actions on match results.
- To investigate any correlations between specific match statistics and overall team performance throughout the dataset period.
- To gain insights into potential predictors of match outcomes and assess the predictive power of statistical models.

# Data exploration and cleaning

The CSV file downloaded from the website contains data for each season of the Spanish La Liga, starting from the 2009/2010 season and spanning 2022/2023 season. Each season's data is structured with various match statistics, including final and half-time scores, team information, and disciplinary actions. The dataset provides a comprehensive overview of match outcomes and related metrics for analysis spanning multiple seasons.

I filtered out qualitative variables and statistics related to betting from the dataset, retaining only the essential match statistics for subsequent analysis.

```{r}
# Read the dataset from the CSV file
football_data <- read_csv("./dataset.csv")
head(football_data)
```

To ensure the integrity of our analysis, we need to clean the data by checking for missing values, duplicate entries, and inconsistencies in data types.

```{r}
# Check for missing values
missing_values <- colSums(is.na(football_data))
missing_values[missing_values > 0]

# Convert necessary columns to appropriate data types
football_data$FTR <- factor(football_data$FTR, levels = c("H", "D", "A"), labels = c("Home Win", "Draw", "Away Win"))

# Summary of the cleaned dataset
summary(football_data)
```

# Univariate Analysis

Distribution of Match Outcomes

```{r}
p1 <- ggplot(football_data, aes(x = FTR)) +
  geom_bar(fill = "lightblue", color = "black") + 
  labs(title = "Distribution of Match Results", x = "Match Outcome", y = "Count") +
  theme_minimal() +
  theme(panel.border = element_rect(color = "black", fill = NA, size = 1))  

p1_interactive <- ggplotly(p1)

p1_interactive
```

Goals Scored Distribution

```{r}
# Home Team Goals
p2 <- ggplot(football_data, aes(x = FTHG)) +
  geom_histogram(aes(text = ..count..), bins = 10, fill = "green", alpha = 0.7, color = "black") +  
  labs(title = "Distribution of Home Team Goals", x = "Goals", y = "Count") +
  theme_minimal() +
  theme(panel.border = element_rect(color = "black", fill = NA, size = 1))  

p2_interactive <- ggplotly(p2, tooltip = "text")

p2_interactive

# Away Team Goals
p3 <- ggplot(football_data, aes(x = FTAG)) +
  geom_histogram(aes(text = ..count..), bins = 10, fill = "red", alpha = 0.7, color = "black") +  
  labs(title = "Distribution of Away Team Goals", x = "Goals", y = "Count") +
  theme_minimal() +
  theme(panel.border = element_rect(color = "black", fill = NA, size = 1)) 

p3_interactive <- ggplotly(p3, tooltip = "text")

p3_interactive
```

Home Advantage
```{r}
# Analyze home advantage
p_home_advantage <- ggplot(football_data, aes(x = FTR, fill = FTR)) +
  geom_bar(aes(text = ..count..), position = "dodge", color = "black") +  
  labs(title = "Home Advantage in Match Outcomes", x = "Match Result", y = "Count") +
  theme_minimal() +
  theme(panel.border = element_rect(color = "black", fill = NA, size = 1))  

ggplotly(p_home_advantage, tooltip = "text")
```

# Bivariate Analysis

Goals vs. Match Outcome
```{r}
# Home Goals vs. Match Outcome
p_home_goals <- ggplot(football_data, aes(x = FTR, y = FTHG)) +
  geom_boxplot(aes(text = paste("Home Goals: ", FTHG)), fill = "lightblue", color = "black") +  
  labs(title = "Home Goals vs. Match Outcome", x = "Match Outcome", y = "Home Goals") +
  theme_minimal() +
  theme(panel.border = element_rect(color = "black", fill = NA, size = 1))

ggplotly(p_home_goals, tooltip = "text")


# Away Goals vs. Match Outcome
p_away_goals <- ggplot(football_data, aes(x = FTR, y = FTAG)) +
  geom_boxplot(aes(text = paste("Away Goals: ", FTAG)), fill = "lightgreen", color = "black") +  
  labs(title = "Away Goals vs. Match Outcome", x = "Match Outcome", y = "Away Goals") +
  theme_minimal() +
  theme(panel.border = element_rect(color = "black", fill = NA, size = 1))


ggplotly(p_away_goals, tooltip = "text")

```


Shots and Match Outcome

We will analyze the relationship between shots and match results by visualizing the number of home and away shots for each match outcome.

```{r}
# Home Team Shots vs. Match Outcome
home_shots_plot <- ggplot(football_data, aes(x = FTR, y = HS)) +
  geom_boxplot(aes(fill = FTR)) +
  labs(title = "Home Team Shots vs. Match Outcome", x = "Full Time Result", y = "Home Team Shots")
ggplotly(home_shots_plot)

# Away Team Shots vs. Match Outcome
away_shots_plot <- ggplot(football_data, aes(x = FTR, y = AS)) +
  geom_boxplot(aes(fill = FTR)) +
  labs(title = "Away Team Shots vs. Match Outcome", x = "Full Time Result", y = "Away Team Shots")
ggplotly(away_shots_plot)

```

# Correlation Matrix

```{r}
numeric_columns <- football_data[, c("FTHG", "FTAG", "HS", "AS", "HST", "AST", "HF", "AF", "HC", "AC", "HY", "AY", "HR", "AR")]

# correlation matrix
cor_matrix <- cor(numeric_columns, use = "complete.obs")

# Correlation Heatmap
heatmap_plot <- plot_ly(
  z = cor_matrix,
  x = colnames(cor_matrix),
  y = colnames(cor_matrix),
  type = "heatmap",
  colors = colorRamp(c("blue", "white", "red")),
  colorbar = list(title = "Correlation")
) %>% layout(
  title = "Correlation Heatmap of Key Match Variables",
  xaxis = list(tickangle = 45),
  yaxis = list(autorange = "reversed")
)

heatmap_plot
```

# Team Data

An interesting list to have in order to manage the data is the list of teams. This is extracted using the unique function as follows:

```{r}
teams <- as.character(unique(football_data[,"HomeTeam"]))
```

```{r}
cat(teams, sep = "\n")
```

# Team Analysis

To begin the analysis, I have decided to start with only one team to simplify the operations. In this case, I have selected **FC Barcelona** as my team to analyze. The information is split into two different dataframes: one for the matches played as the Home Team and the other for the matches played as the Away Team.

```{r}
# Filter Barcelona's matches from the dataset
barcelona_matches <- football_data %>%
  filter(HomeTeam == "Barcelona" | AwayTeam == "Barcelona")

# Separate matches by home and away games
barca_home <- barcelona_matches %>%
  filter(HomeTeam == "Barcelona")

barca_away <- barcelona_matches %>%
  filter(AwayTeam == "Barcelona")
```

Aggregate statistics are calculated for matches where Barcelona played at home and away. This includes total fouls, red/yellow cards, shots, and shots on target.

```{r}
# for home matches
barca_home_summary <- barca_home %>%
  summarize(
    TotalFouls = sum(HF),
    TotalRedCards = sum(HR),
    TotalYellowCards = sum(HY),
    TotalShots = sum(HS),
    TotalShotsOnTarget = sum(HST)
  )

# for away matches
barca_away_summary <- barca_away %>%
  summarize(
    TotalFouls = sum(AF),
    TotalRedCards = sum(AR),
    TotalYellowCards = sum(AY),
    TotalShots = sum(AS),
    TotalShotsOnTarget = sum(AST)
  )

print(barca_home_summary)
print(barca_away_summary)

```

I combined the home and away summaries for easier comparison. Then, the data is reshaped into a long format suitable for plotting, with each statistic represented separately.

```{r}
# Combine summaries and add MatchType information
combined_summary <- bind_rows(
  mutate(barca_home_summary, MatchType = "Home"),
  mutate(barca_away_summary, MatchType = "Away")
)

# Reshape for plotting
combined_summary_long <- pivot_longer(combined_summary, 
                                      cols = c(TotalFouls, TotalRedCards, TotalYellowCards, TotalShots, TotalShotsOnTarget),
                                      names_to = "Statistic",
                                      values_to = "Count")

```

A bar plot is created with custom colors and labels for each statistic. This shows the distribution of fouls, red/yellow cards, and shots across home and away matches.

```{r}
# Define custom color palette for the plot
my_colors <- c("TotalFouls" = "#1f77b4",          
               "TotalRedCards" = "red",           
               "TotalYellowCards" = "yellow",      
               "TotalShots" = "#2ca02c",          
               "TotalShotsOnTarget" = "violet")    

bar_plot <- ggplot(combined_summary_long, aes(x = MatchType, y = Count, fill = Statistic, label = Count)) +
  geom_bar(stat = "identity", position = position_dodge(), color = "black") +
  geom_text(position = position_dodge(width = 0.9), vjust = -0.5, size = 3, 
            aes(group = Statistic), color = "black", fontface = "bold", show.legend = FALSE) +
  labs(title = "Summary of Barcelona Matches",
       y = "Count", x = "Match Type", fill = "Statistic") +
  scale_fill_manual(values = my_colors) +
  theme_minimal() +
  theme(legend.position = "top",
        axis.title.x = element_text(size = 12, face = "bold"),
        axis.title.y = element_text(size = 12, face = "bold"),
        plot.title = element_text(size = 14, face = "bold", hjust = 0.5))

interactive_plot <- ggplotly(bar_plot)
interactive_plot
```

# Input Data preparation (removing multicollinear variables)

## Prepare the Correlation Matrix

```{r}
# Display the structure of football_data
str(football_data)

# Extract numeric variables only
df_corr <- football_data[, sapply(football_data, is.numeric)]

# Create the correlation matrix using Pearson's method
df_corr.cor <- cor(df_corr, method = "pearson")

# Define color palette for heatmap
palette <- colorRampPalette(c("green", "white", "red"))(20)

# Plot correlation heatmap
heatmap(x = df_corr.cor, col = palette, symm = TRUE)

```

The correlation matrix helps in understanding the linear relationships between pairs of numeric variables by presenting a matrix of correlation coefficients.

This heatmap visualizes the strength and direction of linear relationships between numeric variables in football_data:

Green represents positive correlation. Red represents negative correlation. White is near zero, indicating little or no correlation.

## Remove Multicollinear Variables

```{r}
# Select a subset of columns to reduce multicollinearity
df_corr <- df_corr[c(7:16)]

# Recompute the correlation matrix after removing multicollinear variables
df_corr.cor <- cor(df_corr, method = "pearson")

# Plot the updated heatmap
heatmap(x = df_corr.cor, col = palette, symm = TRUE)

```

This step involves identifying and removing multicollinear variables from the dataset. Multicollinearity occurs when independent variables are highly correlated with each other, which can lead to instability and inflated standard errors in regression analysis. By removing multicollinear variables, we streamline the dataset for further analysis, reducing the risk of multicollinearity-related issues and improving the reliability of regression models.

## Prepare the Input Dataset

```{r}
input_data <- football_data[c(1:3, 6:21)]
```

# Split Data into Training and Testing Sets

Now we split the data into a 70% training dataset and a 30% test dataset.

```{r}
# Set seed for reproducibility
set.seed(123)

# Split index creation
index <- createDataPartition(input_data$FTR, p = 0.7, list = FALSE)

train_data <- input_data[index, ]
test_data <- input_data[-index, ]

write.csv(train_data, "./training/training.csv", row.names = FALSE) 
write.csv(test_data, "./testing/test.csv", row.names = FALSE)

```
